Remove instance of `GodToolsProperties` that is not used here.
There are errors that I don't fully understand yet that seem to be tied to the update notifications endpoint.

The primary thing is that the database URL is null when trying to load an Sql2o.  A secondary error is that errbit can't report the error b/c of "too many open files"

I think the errors might be tied to too many instances of the properties file being open at once, and the inclusion of a properties file like this definitely seems out of place