Make the language look-up be non case sensitive
This makes the adding language process more forgiving in case language codes/locales are entered in the wrong case.  I think OneSky still requires them in the right case, but this update makes it so that we can correct it in our database, but the phones can still cache the wrong cased value and still get the right results.  Otherwise we have to release an update to the phones w/ a database update which is dumb, or people have to reinstall the app